Move drawing maze instantly from grid to visualiser
The code to draw a maze as a whole has been moved from the grid to the
visualiser class. This makes it possible to also place the markers for
the entry and exit point.

The visualiser is also a more logical place to house this method as it
is the visualiser that should know how to draw the grid.