Determine entry and exit on longest path in maze
The maze used to always start on an outer cell and take the longest path
from that cell as the exit. This has been changed to generating a maze
from a random cell on the grid, it no longer has to start generating the
maze from on outer edge cell.

After the maze has been constructed there is now a test to find the
longest path for each of the outer cells. Whichever cell has the longest
path through the graph will be set as the entry cell. Its corresponding
exit cell will become the exit of the maze.