Change maze generator backtracking
Updated the backtracking of the maze generator to mimic the method used
for discovering the longest path. It reduces the number of private
properties needed for the class and makes the whole method a bit easier
to read.

This commit also fixes that the first cell to get backtracked was
published twice as a history record. This is no longer done, instead the
event that dispatched for each step taken now has a "state" property.
When it is "backtrack" it means the cell was popped from the stack.