Add web worker implementation for finding longest path
This commit adds a new version of finding the longest path. The new
version uses Web Workers to determine the longest path for multiple
cells at a time.

In order to pass the maze along to the Web Worker the generator has been
given a new method to return a simpler representation of the maze. This
simpler version has no Cell instances, only arrays an objects. This
should help with moving the maze data from the main thread to the Web
Workers.